refactor: moved interfaces to helper file
