+++ packages/auto-label/test/auto-label.ts
@@ -23,8 +23,8 @@ import {resolve} from 'path';
 import fs from 'fs';
 import snapshot from 'snap-shot-it';
 import * as sinon from 'sinon';
-import {handler, DriftRepo, DriftApi} from '../src/auto-label';
-import {autoDetectLabel} from '../src/helper';
+import {handler} from '../src/auto-label';
+import {autoDetectLabel, DriftRepo, DriftApi} from '../src/helper';
 import {logger} from 'gcf-utils';
 nock.disableNetConnect();
 const sandbox = sinon.createSandbox();
